add snapshot tests
