+++ services/QuillLMS/client/app/bundles/Staff/components/adminVerification.tsx/__tests__/approveModal.test.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import { mount } from 'enzyme';
+
+import ApproveModal from '../approveModal';
+
+describe('ApproveModal component', () => {
+  const mockProps = {
+    closeModal: jest.fn(),
+    approve: jest.fn()
+  }
+
+  const container = mount(<ApproveModal {...mockProps} />);
+
+  it('should render ApproveModal', () => {
+    expect(container).toMatchSnapshot();
+  });
+});
